Issue_1146 Removed org/csstudio/display/pace/model/VTypeHelper, renamed org/csstudio/scan/device/VTypeHelper
