+++ services/scan-server/src/main/java/org/csstudio/scan/server/command/SetCommandImpl.java
@@ -19,7 +19,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.csstudio.scan.command.SetCommand;
-import org.csstudio.scan.device.VTypeHelper;
 import org.csstudio.scan.server.MacroContext;
 import org.csstudio.scan.server.ScanCommandImpl;
 import org.csstudio.scan.server.ScanContext;
@@ -27,6 +26,7 @@ import org.csstudio.scan.server.SimulationContext;
 import org.csstudio.scan.server.WriteHelper;
 import org.csstudio.scan.server.device.SimulatedDevice;
 import org.csstudio.scan.server.internal.JythonSupport;
+import org.phoebus.core.vtypes.VTypeHelper;
 import org.phoebus.util.time.TimeDuration;
 
 /** {@link ScanCommandImpl} that sets a device to a value
