+++ services/scan-server/src/main/java/org/csstudio/scan/server/device/PVDevice.java
@@ -25,7 +25,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.logging.Level;
 
 import org.csstudio.scan.device.DeviceInfo;
-import org.csstudio.scan.device.VTypeHelper;
 import org.epics.vtype.Alarm;
 import org.epics.vtype.Time;
 import org.epics.vtype.VByteArray;
@@ -34,6 +33,7 @@ import org.epics.vtype.VType;
 import org.phoebus.pv.PV;
 import org.phoebus.pv.PVPool;
 import org.phoebus.util.time.TimeDuration;
+import org.phoebus.core.vtypes.VTypeHelper;
 
 import io.reactivex.disposables.Disposable;
